CHORE update name to game_ops
