+++ vm/include/game_ops.h
@@ -4,14 +4,16 @@
 
 // function prototypes
 void update_cursor(core_t *core, cursor_t *cursor, int i);
-void update_cycles(core_t **core);
+int update_cycles(core_t **core);
 void update_carry(cursor_t *cursor, int dest_reg);
 int get_reg(core_t *core, cursor_t *cursor, int i);
 int get_dir(core_t *core, cursor_t *cursor, int i);
 int get_label(core_t *core, cursor_t *cursor, int i);
 int	get_ind(core_t *core, cursor_t *cursor, int i);
-void is_alive(core_t *core);
+void add_cycle(int opcode, cursor_t *cursor);
+int is_alive(core_t *core);
+void game_over(core_t *core);
 void print_stats(core_t *core);
-void add_cycle(cursor_t *cursor);
+void print_winner(core_t *core);
 
 #endif
