+++ vm/src/op_add.c
@@ -1,4 +1,4 @@
-#include "../include/op_helper.h"
+#include "../include/game_ops.h"
 
 /* Takes 3 parameters: 3 registers. Add the first to the second, and store
 the result to the third. Modifies the carry. */
