+++ vm/src/op_and.c
@@ -1,4 +1,4 @@
-#include "../include/op_helper.h"
+#include "../include/game_ops.h"
 
 /* Same as add, and sub, but does a binary operation AND between the
 first and the second, and storing the result in the third parameter.
