+++ vm/src/op_ld.c
@@ -1,4 +1,4 @@
-#include "../include/op_helper.h"
+#include "../include/game_ops.h"
 
 /* Takes 2 parameters: loads the first parameter to the second parameter.
 Second parameter is a register. Ex: ld 34, r3 loads the REG_SIZE bytes
