+++ vm/src/op_lfork.c
@@ -1,4 +1,4 @@
-#include "../include/op_helper.h"
+#include "../include/game_ops.h"
 
 /* Same as fork without the %IDX_MOD. */
