+++ vm/src/op_lld.c
@@ -1,4 +1,4 @@
-#include "../include/op_helper.h"
+#include "../include/game_ops.h"
 
 /* Same as ld without the %IDX_MOD. Modifies the carry */
