+++ vm/src/op_st.c
@@ -1,4 +1,4 @@
-#include "../include/op_helper.h"
+#include "../include/game_ops.h"
 
 /* Takes 2 parameters: Stores first parameter (a register) into the second
 parameter (it can be another register or a number). Ex: st r4, 34 stores
